Cleanup:   * Change int to size_t in a few APIs.   * Rename infobar_delegate_count() to infobar_count() in preparation for TabContents owning InfoBars rather than InfoBarDelegates.   * Move some code from PluginInstallerInfoBarDelegate to PluginObserver since it's more related to who's instantiating the infobar(delegate)s.   * Unify InfoBarDelegate behavior by making no delegates auto-add themselves to tabs (callers now all do this explicitly).   * Eliminate unused member TabContentsSSLHelper::SSLAddCertData::handler_.   * De-inline a couple classes.   * Make more functions private.   * Add some consts.   * Change DCHECK() to DCHECK_EQ() where possible.   * Rename the delegates in plugin_observer.cc to have "Delegate" in the names.   * Remove unnecessary qualifiers.   * Simplify.   * Misc. style issues, naming issues, etc.
BUG=none
TEST=none
Review URL: http://codereview.chromium.org/6250172

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@74086 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/notifications/notifications_interactive_uitest.cc
@@ -1,5 +1,5 @@
 FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
-// Copyright (c) 2010 The Chromium Authors. All rights reserved.
+// Copyright (c) 2011 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -62,7 +62,7 @@ TEST_F(NotificationsPermissionTest, FLAKY_TestNoUserGestureInfobar) {
                 "files/notifications/notifications_request_inline.html")));
   WaitUntilTabCount(1);
 
-  int info_bar_count;
+  size_t info_bar_count;
   ASSERT_TRUE(tab->GetInfoBarCount(&info_bar_count));
-  EXPECT_EQ(0, info_bar_count);
+  EXPECT_EQ(0U, info_bar_count);
 }
